+++ example/Monad/Foo.hs
@@ -1,6 +1,5 @@
 module Monad.Foo where
 import Monad.Impl
-import Control.Monad.Fail
 
 data Foo a = Foo a
 
@@ -9,5 +8,3 @@ impl @Monad [t|Foo|]
   $$ #bind   [|\(Foo a) f -> f a|]
   $$ #fmap   [|\f (Foo a) -> Foo (f a)|]
   $$ defaults
-
-