+++ Changelog.md
@@ -2,9 +2,9 @@
 
 ## [0.2.0.0] - 2018-11-26
 ### Added
-- Changelog
 - Optional methods and some `named` utilities to support them
 - Improved example showing heirarchical instance dispatch and optional methods
+- Much improved documentation
 - Exposed `(++)` type family for appending methods
 
 ### Changed
@@ -13,3 +13,4 @@ to support optional methods.
 - Renamed Named parameter application operator from (!) to ($$) to mintain downstream compatibility
 
 ### Removed
+- Deprecated `methodsFor` until `reify` is fixed upstream